refactor: 미 사용 코드 제거
