+++ module-activity-service/src/main/java/com/shop/activityservice/post/service/PostService.java
@@ -4,7 +4,7 @@ import com.shop.activityservice.post.entiity.Post;
 import com.shop.activityservice.post.entiity.PostLike;
 import com.shop.activityservice.post.dto.request.PostWriteRequest;
 import com.shop.activityservice.common.exception.BaseException;
-import com.shop.activityservice.common.exception.ErrorCode;
+import com.shop.activityservice.common.response.ErrorCode;
 import com.shop.activityservice.post.repository.PostLikeRepository;
 import com.shop.activityservice.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
