+++ module-order-service/src/main/java/com/shop/orderservice/common/response/ErrorCode.java
@@ -10,24 +10,6 @@ public enum ErrorCode {
 
     INVALID_REQUEST(HttpStatus.BAD_REQUEST, "Invalid request"),
 
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "User not founded"),
-    INVALID_PASSWORD(HttpStatus.UNAUTHORIZED, "Invalid password"),
-    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "Invalid token"),
-    INVALID_AUTH_CODE(HttpStatus.UNAUTHORIZED, "Invalid email authentication code"),
-
-    POST_NOT_FOUND(HttpStatus.NOT_FOUND, "Post not founded"),
-    COMMENT_NOT_FOUND(HttpStatus.NOT_FOUND, "Comment not founded"),
-
-    DUPLICATED_USER_EMAIL(HttpStatus.CONFLICT, "Duplicated user email"),
-    DUPLICATED_POST_LIKE(HttpStatus.CONFLICT, "User already like the post"),
-    DUPLICATED_COMMENT_LIKE(HttpStatus.CONFLICT, "User already like the comment"),
-
-    ALREADY_USER_FOLLOW(HttpStatus.CONFLICT, "Already user followed"),
-    NOT_FOLLOW_ME(HttpStatus.BAD_REQUEST, "Users can't follow themselves"),
-
-    PRODUCT_NOT_FOUND(HttpStatus.NOT_FOUND, "Product not founded"),
-    DUPLICATED_PRODUCT_TITLE(HttpStatus.CONFLICT, "Duplicated product title"),
-
     ORDER_NOT_FOUND(HttpStatus.NOT_FOUND, "Order infomation not found"),
     ORDER_NOT_AVAILABLE(HttpStatus.BAD_REQUEST, "Not an available time to order"),
